Move AudioGraph::Sink to lib-audio-graph
