+++ libraries/lib-audio-graph/AudioGraphSink.cpp
@@ -1,0 +1,12 @@
+/**********************************************************************
+
+  Audacity: A Digital Audio Editor
+
+  @file AudioGraphSink.cpp
+
+  Paul Licameli
+
+**********************************************************************/
+#include "AudioGraphSink.h"
+
+AudioGraph::Sink::~Sink() = default;
